Added custom discrete colormap class
Class to manage a discrete qualitative colormap. Matplotlib's builtin
discrete colormaps try to normalize the map linearly across the
domain of values. This colormap will assign a color to each given value,
without trying to fill the gaps. Does not extend matplotlib's colormap
class, so it is an external tool used by this plotting module.