Modified new experiment generation
Use seperate generator function to create a new experiment to run.
Makes it easier to require certain parameters for running the experiment
without interfering with recreating the experiment from the database
using the regular constructor. The `new` function will setup the value
iterators in addition to the things managed by the regular constructor.